Define which keys can be used to satisfy a challenge.
This clarifies which keys can be used to satisfy the new ACME challenge
type. Along with the preexisting text that "The CSR MUST include the
public key, attested within the Trust Chain, used by the Requestor to
satisfy the Issuer's challenge.", this also defines which keys can be
included in the eventual certificate -- I made some editorial changes to
highlight this requirement.