Define which keys can be used to satisfy a challenge. (#5)
* Update introductory & concluding sections to match challenge-type spec.

The text of the initial sections is updated to match the new phrasing of
this spec as a new ACME challenge type. I also reflow the text to 80
columns and make a few editorial/typographical changes.

* Define which keys can be used to satisfy a challenge.

This clarifies which keys can be used to satisfy the new ACME challenge
type. Along with the preexisting text that "The CSR MUST include the
public key, attested within the Trust Chain, used by the Requestor to
satisfy the Issuer's challenge.", this also defines which keys can be
included in the eventual certificate -- I made some editorial changes to
highlight this requirement.

---------

Co-authored-by: Giuseppe De Marco <demarcog83@gmail.com>