Remove non-existent response from sequence diagram
The diagram shows the requestor POSTing to the challenge at L326, and
then the issuer responds at L332 indicating validation has begun. The
existing diagram also illustrates a response from issuer to requestor at
L372, but that isn't correct: the requestor's request was previously
responded to at L332. Validation happens asynchronously from the POST to
the challenge, and the requestor learns of its success or failure by
polling the authz.