Fix example acme_requestor JWKS
The example of a JWKS included in an acme_requestor metadata block
should not include `iss`, `sub` or `iat`. The protocol text says that
the JWK set should be published according to OIDF 5.2.1, which describes
metadata keys `signed_jwks_uri`, `jwks_uri` and `jwks`. Only in the
first case are `iss`, `sub` or `iat` necessary. Including them when the
key set is inline in the metadata is just confusing, because they de
facto must have the same `sub` as the enclosing entity configuration.