Allow issuer to provide TAs to requestor
When solving an `openid-federation-01` challenge, the requestor may
provide a trust chain from itself to a trust anchor trusted by the
issuer, to save the issuer the trouble of doing a resolve. However it's
not clear how the requestor is supposed to know what TAs the issuer
trusts.

This commit adds an optional field `trust_anchors` to the challenge
object. The requestor can use that to construct a trust chain that the
issuer will accept.

Additionally, this removes the sentence "The Requestor SHOULD use a
Trust Anchor it has in common with the ACME server."

First, I don't think it's the right requirement. What matters is that
(1) the trust anchor be trusted by the issuer and (2) there be some path
from the requestor to the trust anchor. Those can be true even if the
requestor doesn't trust the TA! To illustrate:

trust_anchor_1           trust_anchor_2
     |________________________|
                  |
            intermediate
                  |
     +------------+-----------+
     |                        |
  issuer                 requestor

If issuer trusts trust_anchor_1 but not _2, and requestor trusts
trust_anchor_2 but not _1, then trust_anchor_1->intermediate->requestor
is a valid trust chain for solving a challenge even if requestor
wouldn't trust it.

Second, it makes more sense as a MUST than a SHOULD. I rewrote it to
make it explicit that the trust chain must end in one of the issuer's
trust anchors.

Resolves #79